% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.pd-package.R
\docType{package}
\name{gsm.pd-package}
\alias{gsm.pd}
\alias{gsm.pd-package}
\title{gsm.pd: Protocol Deviations gsm Plugin}
\description{
Protocol deviations modules for use in the `gsm.app` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openrbqm.github.io/gsm.pd/}
  \item \url{https://github.com/OpenRBQM/gsm.pd}
  \item Report bugs at \url{https://github.com/OpenRBQM/gsm.pd/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
  \item Spencer Childress \email{spencer.childress@gilead.com}
}

}
\keyword{internal}
