% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginPD.R
\name{pluginPD}
\alias{pluginPD}
\title{PD Plugin Definition}
\usage{
pluginPD()
}
\value{
A plugin definition for use in gsm.app.
}
\description{
Plugins for gsm.app are defined by a named \code{list} with information about the
plugin. These can be loaded from a yaml file with \code{\link[gsm.app:plugin_Read]{gsm.app::plugin_Read()}}.
We provide a ready-to-use plugin definition for this PD plugin.
}
\examples{
pluginPD()
}
